OP-1216 Return BigDecimal from the ward-quantity repository queries
Change findQuantityInWardWhereMedical / findQuantityInWardWhereMedicalAndWard to return
BigDecimal instead of Double (the underlying sum of in/out ward quantities is DECIMAL),
and adapt the callers in MedicalStockWardIoOperations — this also removes the lossy
BigDecimal.valueOf(double) conversion when populating MedicalWard.qty.

+++ src/main/java/org/isf/medicalstockward/service/MedicalStockWardIoOperationRepository.java
@@ -44,11 +44,11 @@ public interface MedicalStockWardIoOperationRepository extends JpaRepository<Med
 
 	@Query(value = "select sum(medWard.in_quantity-medWard.out_quantity) from MedicalWard medWard " +
 			"where medWard.id.medical.code=:medical")
-	Double findQuantityInWardWhereMedical(@Param("medical") int medical);
+	BigDecimal findQuantityInWardWhereMedical(@Param("medical") int medical);
 
 	@Query(value = "select sum(medWard.in_quantity-medWard.out_quantity) from MedicalWard medWard " +
 			"where medWard.id.medical.code=:medical and medWard.id.ward.code=:ward")
-	Double findQuantityInWardWhereMedicalAndWard(@Param("medical") int medical, @Param("ward") String ward);
+	BigDecimal findQuantityInWardWhereMedicalAndWard(@Param("medical") int medical, @Param("ward") String ward);
 
 	@Modifying
 	@Query(value = "UPDATE OH_MEDICALDSRWARD SET MDSRWRD_IN_QTI = MDSRWRD_IN_QTI + :quantity WHERE MDSRWRD_WRD_ID_A = :ward AND MDSRWRD_MDSR_ID = :medical AND MDSRWRD_LT_ID_A = :lot", nativeQuery = true)
